@@ -1,7 +1,7 @@
 {
     "name": "hackeresq/laravel-settings",
     "description": "Super simple key/value settings for Laravel 5+",
-    "version": "2.0.1",
+    "version": "3.0.0",
     "keywords": [
         "settings",
         "laravel",
@@ -17,8 +17,7 @@
     },
     "require": {
         "php" : ">=7.0",
-        "laravel/framework": "~5.4.0|~5.5.0|~5.6.0|~5.7.0|~5.8.0|^6.0",
-        "doctrine/dbal": "^2.8"
+        "laravel/framework": "^6.0|^7.0|^8.0"
     },
     "require-dev": {
         "phpunit/phpunit": "4.8.*"
