@@ -2,38 +2,48 @@
 
 namespace hackerESQ\Settings;
 
-use DB;
-use Cache;
+use Illuminate\Support\Arr;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Cache;
 
 class Settings
 {
+    protected string $tenant = '';
 
     /**
      * Get settings from the database
-     * @param string $tenant
      * @return array
      */
-    public function resolveDB($tenant) {
+    public function resolveDB() 
+    {
+        return DB::table('settings')->where('tenant',$this->tenant)->pluck('value', 'key')->toArray();
+    }
 
-        return DB::table('settings')->where('tenant','=',$tenant)->pluck('value', 'key')->toArray();
-        
+    /**
+     * Set the tenant
+     * @param string $tenant (optional)
+     * @return Settings $this
+     */
+    public function tenant($tenant='') 
+    {
+        $this->tenant = $tenant;
+
+        return $this;
     }
 
     /**
      * Get settings from the cache
-     * @param string $tenant
      * @return array
      */
-    public function resolveCache($tenant) {
-
+    public function resolveCache() 
+    {
         if (config('settings.cache')) {
-            return Cache::rememberForever('settings'.$tenant, function () use ($tenant) {
-                return $this->resolveDB($tenant);
+            return Cache::rememberForever('settings'.$this->tenant, function () {
+                return $this->resolveDB();
             });
-        } else {
-            return $this->resolveDB($tenant);
         }
-        
+
+        return $this->resolveDB();
     }
 
     /**
@@ -41,37 +51,29 @@ class Settings
      * @param array $settings
      * @return array
      */
-    public function decryptHandler($settings) {
-
+    public function decryptHandler($settings) 
+    {
         // DO WE NEED TO DECRYPT ANYTHING?
         foreach ($settings as $key => $value) {
             if ( in_array( $key, config('settings.encrypt',[]) ) && !empty($value) ) {
-                array_set($settings, $key, decrypt($settings[$key]));
+                Arr::set($settings, $key, decrypt($settings[$key]));
             }
         }
 
         return $settings;
-
     }
 
     /**
      * Get value of settings by key
-     * @param  string  $key
-     * @param  array  $options
+     * @param string $key 
      * @return mixed string|boolean
      */
-    public function get($key = NULL, $options = [])
+    public function get($key = NULL)
     {
-        // is this multitenant? 
-        $tenant = isset($options['tenant']) ? $options['tenant'] : '';
-
-        $settings = $this->decryptHandler($this->resolveCache($tenant));
+        $settings = $this->decryptHandler($this->resolveCache());
 
         // no key passed, assuming get all settings
-        if ($key == NULL) {
-            
-            return $settings;
-        }
+        if ($key == NULL) return $settings;
         
         // array of keys passed, return those settings only
         if (is_array($key)) {
@@ -88,88 +90,71 @@ class Settings
         } 
 
         return false;
-        
     }
 
     /**
      * Check if a given key exists
-     * @param  string  $key
-     * @param  array  $options
+     * @param string $key
      * @return boolean
      */
-    public function has($key, $options = [])
+    public function has($key)
     {
-
-        // is this multitenant? 
-        $tenant = isset($options['tenant']) ? $options['tenant'] : '';
-        
-        $settings = $this->decryptHandler($this->resolveCache($tenant));
+        $settings = $this->decryptHandler($this->resolveCache());
 
         return array_key_exists($key, $settings);
     }
 
     /**
      * Set value of setting
-     * @param  array  $changes
-     * @param  array  $options
+     * @param array $changes
+     * @param array $options (optional)
      * @return boolean
      */
     public function set($changes, $options = [])
     {
-
-        $force = isset($options['force']) ? $options['force'] : false;
-        $encrypt = isset($options['encrypt']) ? $options['encrypt'] : false;
-
-        // is this multitenant? 
-        $tenant = isset($options['tenant']) ? $options['tenant'] : '';
+        $force = $options['force'] ?? false;
+        $encrypt = $options['encrypt'] ?? false;
 
         // DO WE NEED TO ENCRYPT ANYTHING?
         foreach ($changes as $key => $value) {
             if ( ( in_array($key, config('settings.encrypt') ) || $encrypt ) && !empty($value)) {
-                array_set($changes, $key, encrypt($value));
+                Arr::set($changes, $key, encrypt($value));
             }
         }
 
         // ARE WE FORCING? OR SHOULD WE BE SECURE?
         if (config('settings.force') || $force) {
-
             foreach ($changes as $key => $value) {
-
-                DB::table('settings')->where([
-                    ['key', '=', $key],
-                    ['tenant', '=', $tenant]
-                    ])->delete();    
-
-                DB::table('settings')->insert([
+                DB::table('settings')->updateOrInsert([
+                    'key'=>$key,
+                    'tenant'=>$this->tenant
+                ],
+                [
                     'key'=>$key,
                     'value'=>$value,
-                    'tenant'=>$tenant
+                    'tenant'=>$this->tenant
                 ]); 
             }
-
         } else {
-
-            $settings = $this->resolveCache($tenant);
+            $settings = $this->resolveCache();
 
             // array_only() - will return only the specified key-value pairs from the given array
                 // array_keys() - will return all the keys or a subset of the keys of an array
                     // this passes array_keys() to array_only() to give current/valid settings only
                         //checks and see if passed settings are  valid options
-            foreach (array_only($changes, array_keys($settings)) as $key => $value) {
+            foreach (Arr::only($changes, array_keys($settings)) as $key => $value) {
                 DB::table('settings')->where([
                     ['key', '=', $key],
-                    ['tenant', '=', $tenant]
+                    ['tenant', '=', $this->tenant]
                 ])->update(['value'=>$value]); 
             }
         }
 
+        // clear cache
         if (config('settings.cache')) {
-            Cache::forget('settings'.$tenant);
+            Cache::forget('settings'.$this->tenant);
         }
 
         return true;
-
     }
-
 }
-
